Add files via upload
1.	Introduction:
The proposed system is about a Doctors and patients appointment booking system. This System will help the users to book appointments with the doctors. The system will store information about doctors and users. The admin will have access to approved doctors and modify them. The system will also check the validation for specific date and time of appointment. The documentation of this system covers initial study, feasibility study, analysis, design, implementation and testing. I have used some applications and programming languages to complete this project.

Objectives:
•	Storing all information of Doctors into the database for managing the system.
•	Storing information of users into the database.
•	Storing Booking information into database. It will help to validate the appointment.
•	Admin approval to represent Doctors information to the users.
•	Admin disallow Doctors information to be seen on the interface.
•	Managed Appointments booking between user and doctors. Selecting Booking bookings dates and times.
•	Checking validation of availability for appointments.
•	Sending mail after the confirmation of appointment.